Migrate libraries/recentemojis into libraries/emoji
