+++ libraries/emoji/impl/build.gradle.kts
@@ -1,0 +1,41 @@
+/*
+ * Copyright (c) 2026 Element Creations Ltd.
+ *
+ * SPDX-License-Identifier: AGPL-3.0-only OR LicenseRef-Element-Commercial.
+ * Please see LICENSE files in the repository root for full details.
+ */
+
+import extension.setupDependencyInjection
+import extension.testCommonDependencies
+
+plugins {
+    id("io.element.android-compose-library")
+}
+
+android {
+    namespace = "io.element.android.libraries.emoji.impl"
+    testOptions {
+        unitTests {
+            isIncludeAndroidResources = true
+        }
+    }
+}
+
+setupDependencyInjection()
+
+dependencies {
+    api(projects.libraries.emoji.api)
+    implementation(projects.libraries.architecture)
+    implementation(projects.libraries.core)
+    implementation(projects.libraries.designsystem)
+    implementation(projects.libraries.di)
+    implementation(projects.libraries.matrix.api)
+    implementation(projects.libraries.uiStrings)
+
+    implementation(libs.kotlinx.collections.immutable)
+    implementation(libs.matrix.emojibase.bindings)
+
+    testImplementation(projects.libraries.emoji.test)
+    testImplementation(projects.libraries.matrix.test)
+    testCommonDependencies(libs)
+}
