--- libraries/recentemojis/api/src/main/kotlin/io/element/android/libraries/recentemojis/api/GetRecentEmojis.kt
@@ -1,18 +1,0 @@
-/*
- * Copyright (c) 2025 Element Creations Ltd.
- * Copyright 2025 New Vector Ltd.
- *
- * SPDX-License-Identifier: AGPL-3.0-only OR LicenseRef-Element-Commercial.
- * Please see LICENSE files in the repository root for full details.
- */
-
-package io.element.android.libraries.recentemojis.api
-
-import kotlinx.collections.immutable.ImmutableList
-
-/**
- * Returns the list of recently used emojis for reactions.
- */
-fun interface GetRecentEmojis {
-    suspend operator fun invoke(): Result<ImmutableList<String>>
-}
