--- libraries/recentemojis/impl/build.gradle.kts
@@ -1,37 +1,0 @@
-/*
- * Copyright (c) 2025 Element Creations Ltd.
- * Copyright 2025 New Vector Ltd.
- *
- * SPDX-License-Identifier: AGPL-3.0-only OR LicenseRef-Element-Commercial.
- * Please see LICENSE files in the repository root for full details.
- */
-
-import extension.setupDependencyInjection
-
-plugins {
-    id("io.element.android-compose-library")
-    id("kotlin-parcelize")
-}
-
-android {
-    namespace = "io.element.android.libraries.recentemojis.impl"
-}
-
-setupDependencyInjection()
-
-dependencies {
-    api(projects.libraries.recentemojis.api)
-    implementation(projects.libraries.architecture)
-    implementation(projects.libraries.matrix.api)
-    implementation(libs.kotlinx.collections.immutable)
-    implementation(libs.matrix.emojibase.bindings)
-
-    testImplementation(projects.libraries.recentemojis.test)
-    testImplementation(libs.test.junit)
-    testImplementation(libs.coroutines.test)
-    testImplementation(libs.molecule.runtime)
-    testImplementation(libs.test.truth)
-    testImplementation(libs.test.turbine)
-    testImplementation(projects.libraries.matrix.test)
-    testImplementation(projects.tests.testutils)
-}
