--- libraries/recentemojis/impl/src/main/kotlin/io/element/android/libraries/recentemojis/impl/DefaultAddRecentEmoji.kt
@@ -1,26 +1,0 @@
-/*
- * Copyright (c) 2025 Element Creations Ltd.
- * Copyright 2025 New Vector Ltd.
- *
- * SPDX-License-Identifier: AGPL-3.0-only OR LicenseRef-Element-Commercial.
- * Please see LICENSE files in the repository root for full details.
- */
-
-package io.element.android.libraries.recentemojis.impl
-
-import dev.zacsweers.metro.ContributesBinding
-import io.element.android.libraries.core.coroutine.CoroutineDispatchers
-import io.element.android.libraries.di.SessionScope
-import io.element.android.libraries.matrix.api.MatrixClient
-import io.element.android.libraries.recentemojis.api.AddRecentEmoji
-import kotlinx.coroutines.withContext
-
-@ContributesBinding(SessionScope::class)
-class DefaultAddRecentEmoji(
-    private val client: MatrixClient,
-    private val dispatchers: CoroutineDispatchers,
-) : AddRecentEmoji {
-    override suspend operator fun invoke(emoji: String): Result<Unit> = withContext(dispatchers.io) {
-        client.addRecentEmoji(emoji)
-    }
-}
