--- libraries/recentemojis/test/build.gradle.kts
@@ -1,26 +1,0 @@
-/*
- * Copyright (c) 2025 Element Creations Ltd.
- * Copyright 2023-2025 New Vector Ltd.
- *
- * SPDX-License-Identifier: AGPL-3.0-only OR LicenseRef-Element-Commercial.
- * Please see LICENSE files in the repository root for full details.
- */
-
-plugins {
-    id("io.element.android-library")
-}
-
-android {
-    namespace = "io.element.android.libraries.recentemojis.test"
-}
-
-dependencies {
-    api(projects.libraries.matrix.api)
-    api(libs.coroutines.core)
-
-    implementation(libs.kotlinx.collections.immutable)
-    implementation(libs.coroutines.test)
-    implementation(projects.tests.testutils)
-    implementation(projects.libraries.recentemojis.api)
-    implementation(libs.matrix.emojibase.bindings)
-}
